Adapt to the new project structure, run prettier
